fix(broker): remove deadcode
