+++ deploy/helm/demarkus-broker/Chart.yaml
@@ -3,7 +3,7 @@ name: demarkus-broker
 description: |
   OIDC token broker for demarkus. Exchanges a verified IdP identity for one
   or more demarkus tokens, writing token hashes into per-world Kubernetes
-  Secrets and tracking ownership in a broker-namespace issuances Secret.
+  Secrets.
   Multi-replica HA, leader-elected expiry/drift sweeper, per-subject and
   per-IP rate limiting.
 type: application
