+++ deploy/helm/demarkus-broker/templates/rbac-broker-ns.yaml
@@ -16,12 +16,12 @@
      get/update pinned to the configured lease name; create gets
      its own rule (namespace-wide).
 
-   - secrets (issuances + refresh-tokens): chart-seeded at install
-     time (secret-issuances.yaml + secret-refresh-tokens.yaml), so
-     the broker only needs get/update — never create. The
-     resource-policy: keep annotation on both means helm upgrade
-     never re-templates them, so the broker SA's update verb is
-     the only path that mutates them after install.
+   - secrets (refresh-tokens): chart-seeded at install time
+     (secret-refresh-tokens.yaml), so the broker only needs
+     get/update — never create. The resource-policy: keep
+     annotation means helm upgrade never re-templates it, so the
+     broker SA's update verb is the only path that mutates it
+     after install.
 
    - secrets (per-world write-token): the broker provisions these
      lazily on the first write to each configured world. get/update
@@ -49,7 +49,6 @@ rules:
   - apiGroups: [""]
     resources: ["secrets"]
     resourceNames:
-      - {{ include "demarkus-broker.issuancesSecretName" . | quote }}
       - {{ include "demarkus-broker.refreshTokensSecretName" . | quote }}
     verbs: ["get", "update"]
   {{- if .Values.worlds }}
