+++ deploy/helm/demarkus-broker/templates/secret-config.yaml
@@ -78,7 +78,6 @@ stringData:
       cookieKey: {{ $cookieKey | quote }}
       stateTTL: {{ .Values.server.stateTTL | quote }}
       brokerNamespace: {{ include "demarkus-broker.brokerNamespace" . | quote }}
-      issuancesSecret: {{ include "demarkus-broker.issuancesSecretName" . | quote }}
       refreshTokensSecret: {{ include "demarkus-broker.refreshTokensSecretName" . | quote }}
       refreshTokenTTL: {{ .Values.server.refreshTokenTTL | quote }}
       idTokenTTL: {{ .Values.server.idTokenTTL | quote }}
@@ -96,8 +95,6 @@ stringData:
           certFile: {{ printf "%s/tls.crt" (include "demarkus-broker.mcpTLSMountPath" $) | quote }}
           keyFile: {{ printf "%s/tls.key" (include "demarkus-broker.mcpTLSMountPath" $) | quote }}
         {{- end }}
-        sessionMaxIdle: {{ .Values.server.mcp.sessionMaxIdle | quote }}
-        maxSessions: {{ .Values.server.mcp.maxSessions }}
         firstMintMaxAttempts: {{ .Values.server.mcp.firstMintMaxAttempts }}
         firstMintInitialBackoff: {{ .Values.server.mcp.firstMintInitialBackoff | quote }}
         firstMintMaxBackoff: {{ .Values.server.mcp.firstMintMaxBackoff | quote }}
