--- deploy/helm/demarkus-broker/templates/secret-issuances.yaml
@@ -1,44 +1,0 @@
-{{/*
-Issuances Secret. The broker writes the label -> {email, world, paths,
-operations, issued_at, expires} map here at runtime; this template only
-seeds an empty Secret on first install so the broker has a target to
-write to.
-
-Two helm-correctness invariants:
-
-  1. `helm.sh/resource-policy: keep` — once created, helm uninstall does
-     NOT delete this Secret. Reinstalling the chart picks up the same
-     issuance state, so users keep their tokens across chart lifecycle
-     events.
-
-  2. `lookup`-skip — on `helm upgrade`, the lookup branch returns the
-     existing live Secret, so the template renders nothing. helm sees
-     the resource was removed from the rendered manifest and would
-     normally garbage-collect it; the resource-policy: keep annotation
-     on the LIVE Secret prevents that. Net effect: chart updates never
-     touch the broker's runtime state.
-
-  `lookup` returns nil during `helm template` (no cluster context), so
-  offline renders always emit the empty Secret. helm-unittest assertions
-  see the empty Secret and can verify the resource-policy annotation.
-*/}}
-{{- /* The issuances Secret lives in server.brokerNamespace, which the
-       broker reads at runtime to locate its own issuance state. Default
-       resolves to .Release.Namespace via the helper; an explicit override
-       puts the Secret in a different namespace and requires matching RBAC
-       (shipped in the RBAC slice). Both the lookup and the rendered
-       metadata.namespace use the helper so they always agree. */ -}}
-{{- $brokerNs := include "demarkus-broker.brokerNamespace" . -}}
-{{- if not (lookup "v1" "Secret" $brokerNs (include "demarkus-broker.issuancesSecretName" .)) }}
-apiVersion: v1
-kind: Secret
-metadata:
-  name: {{ include "demarkus-broker.issuancesSecretName" . }}
-  namespace: {{ $brokerNs }}
-  labels:
-    {{- include "demarkus-broker.labels" . | nindent 4 }}
-  annotations:
-    helm.sh/resource-policy: keep
-type: Opaque
-data: {}
-{{- end }}
