+++ deploy/helm/demarkus-broker/templates/secret-refresh-tokens.yaml
@@ -1,8 +1,8 @@
 {{/*
 Refresh-tokens Secret. The broker writes the sha256(refresh_token) ->
 record map here at runtime; this template only seeds an empty Secret
-on first install so the broker has a target to write to. Same
-helm-correctness invariants as secret-issuances.yaml:
+on first install so the broker has a target to write to. Two
+helm-correctness invariants:
 
   1. `helm.sh/resource-policy: keep` — once created, helm uninstall
      does NOT delete this Secret. Reinstalling the chart picks up
@@ -21,9 +21,7 @@ helm-correctness invariants as secret-issuances.yaml:
   assertions see the empty Secret and can verify the resource-
   policy annotation.
 
-PR4 (universe onboarding §6). Mirrors the issuances Secret on
-purpose so operators get the same audit experience for both the
-world-token and refresh-token state.
+PR4 (universe onboarding §6).
 */}}
 {{- $brokerNs := include "demarkus-broker.brokerNamespace" . -}}
 {{- if not (lookup "v1" "Secret" $brokerNs (include "demarkus-broker.refreshTokensSecretName" .)) }}
