--- deploy/helm/demarkus-broker/tests/secret-issuances_test.yaml
@@ -1,46 +1,0 @@
-suite: secret-issuances
-templates:
-  - secret-issuances.yaml
-release:
-  namespace: broker-ns
-set:
-  oidc.issuer: https://accounts.example.com
-  oidc.clientID: broker-app
-  oidc.clientSecret: shh
-  oidc.redirectURL: https://broker.example.com/auth/callback
-tests:
-  - it: issuances Secret renders an empty Opaque Secret on offline template
-    asserts:
-      - hasDocuments:
-          count: 1
-      - isKind:
-          of: Secret
-      - equal:
-          path: metadata.name
-          value: RELEASE-NAME-demarkus-broker-issuances
-      - equal:
-          path: type
-          value: Opaque
-      - equal:
-          path: data
-          value: {}
-
-  - it: helm.sh/resource-policy keep annotation pinned so helm upgrade/uninstall do not wipe broker state
-    asserts:
-      - equal:
-          path: metadata.annotations["helm.sh/resource-policy"]
-          value: keep
-
-  - it: metadata.namespace tracks demarkus-broker.brokerNamespace helper
-    asserts:
-      - equal:
-          path: metadata.namespace
-          value: broker-ns
-
-  - it: server.brokerNamespace override moves the issuances Secret to that namespace
-    set:
-      server.brokerNamespace: alt-broker-ns
-    asserts:
-      - equal:
-          path: metadata.namespace
-          value: alt-broker-ns
