+++ tools/demarkus-broker/internal/broker/labels.go
@@ -1,9 +1,10 @@
 // Package broker implements the demarkus OIDC token broker: a small HTTP
-// service that exchanges an OIDC identity for one or more demarkus tokens,
-// writing token hashes into per-world Kubernetes Secrets while keeping its
-// own issuance state in a separate broker-namespace Secret. World servers
-// stay identity-blind; the broker is the only component that knows who
-// owns which token.
+// service that gates demarkus world access behind SSO. It provisions and
+// holds one long-lived write token per world (hashes land in the world's
+// Kubernetes Secret) and serves the MCP gateway; reads dispatch open with
+// an empty bearer, writes are gated by WorldConfig.Allow. World servers
+// stay identity-blind; the broker is the only component that knows who is
+// behind a given call.
 package broker
 
 import (
@@ -12,16 +13,16 @@ import (
 	"fmt"
 )
 
-// LabelPrefix is the namespace prefix the broker stamps on every label it
-// mints. Admin-minted labels (via the demarkus-token CLI) use any prefix
-// the admin chooses; the broker only operates on labels it created itself,
-// which it identifies by this prefix plus the issuances Secret.
+// LabelPrefix is the prefix the broker stamps on every token label it
+// generates, distinguishing broker-generated labels from admin-minted
+// ones (via the demarkus-token CLI), which use any prefix the admin
+// chooses.
 const LabelPrefix = "usr_"
 
 // NewLabel returns an opaque token label of the form "usr_<8 hex chars>"
-// (4 bytes of entropy). Collision space is 2^32, which is fine because the
-// issuer rejects duplicates on Secret write (ErrLabelExists) and the
-// caller retries on collision.
+// (4 bytes of entropy). Collision space is 2^32; callers that persist a
+// label are expected to reject duplicates on Secret write and retry on
+// collision.
 func NewLabel() (string, error) {
 	var b [4]byte
 	if _, err := rand.Read(b[:]); err != nil {
