+++ tools/go.mod
@@ -13,7 +13,6 @@ require (
 	github.com/latebit/demarkus/protocol v0.0.0
 	github.com/mark3labs/mcp-go v0.44.0
 	golang.org/x/oauth2 v0.36.0
-	golang.org/x/sync v0.20.0
 	golang.org/x/time v0.14.0
 	gopkg.in/yaml.v3 v3.0.1
 	k8s.io/api v0.36.0
